Rename *Service* namespace to *Container*
